fix OSX compile error
+ needs to have explicit cast from float to int in these array initializer lists

+++ SpectrumSelection.cpp
@@ -160,8 +160,8 @@ bool SpectrumSelection::Draw(IGraphics* pGraphics)
 
 void SpectrumSelection::DrawHandle(IGraphics* pGraphics, const IRECT& handle)
 {
-	int x[4] = { handle.MW(), handle.R, handle.MW(), handle.L };
-	int y[4] = { handle.T, handle.MH(), handle.B, handle.MH() };
+	int x[4] = { (int)handle.MW(), handle.R, (int)handle.MW(), handle.L };
+	int y[4] = { handle.T, (int)handle.MH(), handle.B, (int)handle.MH() };
 
 	//pGraphics->DrawLine(&handleColor, x[0], y[0], x[1], y[1], 0, true);
 	//pGraphics->DrawLine(&handleColor, x[1], y[1], x[2], y[2], 0, true);
